% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_package_functions.R
\name{get.extrema.OR}
\alias{get.extrema.OR}
\alias{get.extrema.RR}
\title{Obtain the point estimate intervals of counterfactual
means}
\usage{
get.extrema.OR(A, Y, gamma = 0, fitted.probs)

get.extrema.RR(A, Y, gamma = 0, fitted.probs)
}
\arguments{
\item{A}{treatment indicator}

\item{Y}{outcome}

\item{gamma}{sensitivity parameter (log odds ratio)}

\item{fitted.probs}{fitted (generalized) propensity scores}
}
\value{
The partially identified interval of the IPW estimator.

The partially identified interval of the IPW estimator.
}
\description{
Obtain the partially identified point estimate intervals of counterfactual
means under the Odds Ratio Framework
}
\section{Functions}{
\itemize{
\item \code{get.extrema.RR}: Obtain the partially identified point estimate
intervals of counterfactual means under the Risk Ratio Framework
}}

