% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_package_functions.R
\name{gps.estimate}
\alias{gps.estimate}
\title{Estimate the generalized propensity scores}
\usage{
gps.estimate(data, A_name, gps.formula)
}
\arguments{
\item{data}{A \code{\link[base:data.frame]{data.frame}} containing all variables required for the analysis}

\item{A_name}{the name of the treatment variable in the data frame}

\item{gps.formula}{a formula object that specifies the regression model for GPS}
}
\value{
A matrix of the estimates GPS
}
\description{
Estimate the generalized propensity scores using multinomial logistic
regression
}
