% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_package_functions.R
\name{outreg}
\alias{outreg}
\title{Fit the outcome regression model}
\usage{
outreg(data, outreg.formula, outreg.family = "gaussian")
}
\arguments{
\item{data}{A \code{\link[base:data.frame]{data.frame}} containing all variables required for the analysis}

\item{outreg.formula}{an object of class \code{\link[stats:formula]{formula}} that specifies the regression model for the outcome regression}

\item{outreg.family}{a description of the error distribution and link function to be used in the model. (See \code{\link[stats:family]{family}} for details of family functions.)}
}
\value{
The fitted outcome regression model object
}
\description{
Fit the outcome regression model
}
