% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_package_functions.R
\name{extrema.os.unified}
\alias{extrema.os.unified}
\alias{bootsens.os.unified}
\title{Sensitivity analysis for observational studies}
\usage{
extrema.os.unified(
  data,
  A_name,
  Y_name,
  gps.formula,
  contrast,
  gamma = 0,
  AIPW = FALSE,
  method = "RR",
  outreg.formula = NULL,
  outreg.family = "gaussian"
)

bootsens.os.unified(
  data,
  A_name,
  Y_name,
  gps.formula,
  contrast,
  gamma = 0,
  AIPW = FALSE,
  method = "RR",
  outreg.formula = NULL,
  outreg.family = "gaussian",
  alpha = 0.1,
  parallel = FALSE,
  B = 1000
)
}
\arguments{
\item{data}{A \code{\link[base:data.frame]{data.frame}} containing all variables required for the analysis}

\item{A_name}{the name of the treatment variable in the data frame}

\item{Y_name}{the name of the outcome variable in the data frame}

\item{gps.formula}{an object of class \code{\link[stats:formula]{formula}} that specifies the regression model for GPS}

\item{contrast}{the linear contrast to define the target estimand}

\item{gamma}{sensitivity parameter (log odds ratio or log risk ratio)}

\item{AIPW}{should the doubly robust AIPW estimator be implemented? (TRUE or FALSE)}

\item{method}{The sensitivity analysis framework to be used ("OR" or "RR")}

\item{outreg.formula}{an object of class \code{\link[stats:formula]{formula}} that specifies the regression model for the outcome regression}

\item{outreg.family}{a description of the error distribution and link function to be used in the model. (See \code{\link[stats:family]{family}} for details of family functions.)}

\item{alpha}{Significance level}

\item{parallel}{Should parallel computing be used? (TRUE or FALSE)}

\item{B}{Number of Bootstrap resamples.}
}
\value{
The partially identified interval of the target estimand

A (1 - alpha) confidence interval
}
\description{
Sensitivity analysis for observational studies
}
\section{Functions}{
\itemize{
\item \code{bootsens.os.unified}: Obtain a (1 - alpha) confidence interval under
the specified sensitivity analysis framework
}}

\examples{
\dontrun{
# sensitivity analysis using the mtcars dataset
# Let mpg be the outcome and cyl be the treatment with three levels

# sensitivity analysis for pairwise ATE between cylinder type 4 and 8

# gamma = 0 (No unmeasured confounding)
## point estimate interval
extrema.os.unified(data = mtcars, A_name = "cyl", Y_name = "mpg",
                   gps.formula = cyl ~ disp + hp + drat + wt, gamma = 0,
                   contrast = c(1, 0, -1), method = "RR")

## 90\% Confidence interval
bootsens.os.unified(data = mtcars, A_name = "cyl", Y_name = "mpg",
                    gps.formula = cyl ~ disp + hp + drat + wt, gamma = 0,
                    contrast = c(1, 0, -1), parallel = TRUE, alpha = 0.1)

# gamma = 0.5
## point estimate interval
extrema.os.unified(data = mtcars, A_name = "cyl", Y_name = "mpg",
                   gps.formula = cyl ~ disp + hp + drat + wt, gamma = 0.5,
                   contrast = c(1, 0, -1), method = "RR")

## 90\% Confidence interval
bootsens.os.unified(data = mtcars, A_name = "cyl", Y_name = "mpg",
                    gps.formula = cyl ~ disp + hp + drat + wt, gamma = 0.5,
                    contrast = c(1, 0, -1), parallel = TRUE, alpha = 0.1)
}
\dontrun{
# sensitivity analysis using the mtcars dataset
# Let mpg be the outcome and cyl be the treatment with three levels

# sensitivity analysis for pairwise ATE between cylinder type 4 and 8

# gamma = 0 (No unmeasured confounding)
## point estimate interval
extrema.os.unified(data = mtcars, A_name = "cyl", Y_name = "mpg",
                   gps.formula = cyl ~ disp + hp + drat + wt, gamma = 0,
                   contrast = c(1, 0, -1), method = "RR")

## 90\% Confidence interval
bootsens.os.unified(data = mtcars, A_name = "cyl", Y_name = "mpg",
                    gps.formula = cyl ~ disp + hp + drat + wt, gamma = 0,
                    contrast = c(1, 0, -1), parallel = TRUE, alpha = 0.1)

# gamma = 0.5
## point estimate interval
extrema.os.unified(data = mtcars, A_name = "cyl", Y_name = "mpg",
                   gps.formula = cyl ~ disp + hp + drat + wt, gamma = 0.5,
                   contrast = c(1, 0, -1), method = "RR")

## 90\% Confidence interval
bootsens.os.unified(data = mtcars, A_name = "cyl", Y_name = "mpg",
                    gps.formula = cyl ~ disp + hp + drat + wt, gamma = 0.5,
                    contrast = c(1, 0, -1), parallel = TRUE, alpha = 0.1)
}
}
